pyproject.toml
[project]
name = "hfung4_dbr_ml_utils"
description = "hfung4 common utils for Databricks ML projects"
requires-python = ">=3.11, <3.12"
dependencies = [
    "pandas<=2.1.4",
    "databricks-sdk>=0.46.0",
    "databricks-connect>=15.4.1, <16",
]

dynamic = ['version']

[project.optional-dependencies]
dev = ["ipykernel>=6.29.5, <7", "pip>=24.2", "pre-commit>=4.1.0, <5"]
test = ["pytest>=8.3.5", "pytest-cov>=5.0.0", "pre-commit>=4.1.0, <5"]


[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["hfung4_dbr_ml_utils*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
testpaths = [ "tests" ]
pythonpath = [ "." ]


[tool.ruff]
line-length = 120
select = [
    "F", # pyflakes rules
    "E", # pycodestyle error rules
    "W", # pycodestyle warning rules
    "B", # flake8-bugbear rules
    "I", # isort rules
]
lint.ignore = [
    "D107", # Missing docstring in `__init__`
    "D203", # incorrect-blank-line-before-class
    "D213", # multi-line-summary-second-line
    "E501", # line-too-long
    "C901", # too complex
    "ERA001", # commented out code
    "D100", # missing docstring
    "ERA001" # commented out code
]

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true

